Ui. Implemented basic navigation.
