+++ app/build.gradle
@@ -59,5 +59,10 @@ dependencies {
     androidTestImplementation "androidx.compose.ui:ui-test-junit4:$compose_version"
     debugImplementation "androidx.compose.ui:ui-tooling:$compose_version"
     debugImplementation "androidx.compose.ui:ui-test-manifest:$compose_version"
+
+    //Image Loading
     implementation("io.coil-kt:coil-compose:2.2.1")
+
+    //Navigation
+    implementation("androidx.navigation:navigation-compose:2.5.2")
 }
