+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.TestProject"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".root.RootActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.TestProject">
