+++ app/src/main/java/com/belyakov/testproject/base/presentation/navigation/TestNewsDestination.kt
@@ -1,0 +1,5 @@
+package com.belyakov.testproject.base.presentation.navigation
+
+open class TestNewsDestination(
+    val route: String
+)
