--- app/src/main/java/com/belyakov/testproject/detail/presentation/NewsDetailModel.kt
@@ -1,9 +1,0 @@
-package com.belyakov.testproject.detail.presentation
-
-import androidx.lifecycle.ViewModel
-
-class NewsDetailModel : ViewModel() {
-    fun onNewsClick() {
-
-    }
-}
