--- app/src/main/java/com/belyakov/testproject/main/presentation/MainViewModel.kt
@@ -1,9 +1,0 @@
-package com.belyakov.testproject.main.presentation
-
-import androidx.lifecycle.ViewModel
-
-class MainViewModel : ViewModel() {
-    fun onNewsClick() {
-
-    }
-}
