+++ app/src/main/java/com/belyakov/testproject/newsdetail/presentation/NewsDetailViewModel.kt
@@ -1,0 +1,9 @@
+package com.belyakov.testproject.newsdetail.presentation
+
+import androidx.lifecycle.ViewModel
+
+class NewsDetailViewModel : ViewModel() {
+    fun onNewsClick() {
+
+    }
+}
