+++ app/src/main/java/com/belyakov/testproject/newsdetail/presentation/composable/NewsDetailedTop.kt
@@ -1,64 +1,59 @@
-package com.belyakov.testproject.main.presentation.composable
+package com.belyakov.testproject.newsdetail.presentation.composable
 
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Card
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import com.belyakov.testproject.base.presentation.theme.Cyan100
 import com.belyakov.testproject.base.presentation.theme.Gray500
-import com.belyakov.testproject.main.presentation.model.NewsUiModel
+import com.belyakov.testproject.newslist.presentation.composable.getFakeNewsModel
+import com.belyakov.testproject.newslist.presentation.model.NewsUiModel
 
 @Composable
-fun NewsItem(
+fun NewsDetailedTop(
     newsModel: NewsUiModel,
-    onNewsClick: () -> Unit,
     modifier: Modifier = Modifier
 ) {
-    val itemHeight = 120.dp
-
-    Card(
-        elevation = 4.dp,
-        shape = RoundedCornerShape(12.dp),
-    ) {
-        Row(
+    Box(modifier = modifier.heightIn(min = 360.dp)) {
+        AsyncImage(
+            model = newsModel.imageUrl,
+            contentScale = ContentScale.Crop,
+            contentDescription = null,
             modifier = Modifier
+                .height(300.dp)
                 .fillMaxWidth()
-                .height(itemHeight)
-                .clickable { onNewsClick() }
+        )
+        Card(
+            elevation = 4.dp,
+            shape = RoundedCornerShape(12.dp),
+            modifier = Modifier
+                .align(Alignment.BottomCenter)
+                .padding(horizontal = 12.dp)
+                .heightIn(min = 120.dp)
+                .fillMaxWidth()
         ) {
-            AsyncImage(
-                model = newsModel.imageUrl,
-                contentScale = ContentScale.Crop,
-                modifier = Modifier.size(itemHeight),
-                contentDescription = null
-            )
             Column(
                 verticalArrangement = Arrangement.SpaceBetween,
                 modifier = Modifier
-                    .fillMaxWidth()
-                    .fillMaxHeight()
                     .padding(12.dp)
             ) {
                 Text(
                     text = newsModel.title,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis,
                     style = MaterialTheme.typography.h6
                 )
                 Row(
@@ -83,22 +78,11 @@ fun NewsItem(
     }
 }
 
-@Preview(locale = "en")
+@Preview
 @Composable
-fun NewsItemPreview() {
-    NewsItem(
-        newsModel = getFakeNewsModel(),
+fun NewsDetailedTopPreview() {
+    NewsDetailedTop(
+        getFakeNewsModel(0),
         modifier = Modifier,
-        onNewsClick = {}
-    )
-}
-
-@Preview(locale = "ar")
-@Composable
-fun NewsItemRTLPreview() {
-    NewsItem(
-        newsModel = getFakeNewsModel(),
-        modifier = Modifier,
-        onNewsClick = {}
     )
 }
