+++ app/src/main/java/com/belyakov/testproject/newslist/presentation/NewsListViewModel.kt
@@ -1,0 +1,9 @@
+package com.belyakov.testproject.newslist.presentation
+
+import androidx.lifecycle.ViewModel
+
+class NewsListViewModel : ViewModel() {
+    fun onNewsClick() {
+
+    }
+}
