+++ app/src/main/java/com/belyakov/testproject/root/RootActivity.kt
@@ -1,0 +1,14 @@
+package com.belyakov.testproject.root
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+
+class RootActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            TestNewsApp()
+        }
+    }
+}
